Remove unnecessary webkit2gtk-4.0 and libsoup-2.4
Building webkit2gtk-4.0 in our manifest requires machine resources, time to
build, and frequent update when upstream release for security reasons,
which is just burden for us.

The GNOME runtime comes with webkit2gtk-4.1, and our manifest can builds
successfully with it and it seems to work as expected for us, so use it instead.