Update Poedit to 3.6.3 and update dependencies (#58)
* Update POEdit to 3.6.3

* Fix 'Legacy windowing system' warning on Flathub

* Update lucene++ to 3.0.9

* Update boost to 1.84.0

* Update cpprestsdk to v2.10.19

* Update gettext to 0.25

* Update enchant to v2.8.4

* Remove unnecessary webkit2gtk-4.0 and libsoup-2.4

Building webkit2gtk-4.0 in our manifest requires machine resources, time to
build, and frequent update when upstream release for security reasons,
which is just burden for us.

The GNOME runtime comes with webkit2gtk-4.1, and our manifest can builds
successfully with it and it seems to work as expected for us, so use it instead.

* Remove unnecessary enchant

This is also in the GNOME runtime

* Remove unnecessary unifdef

Because this was added as a dependency of webkit2gtk-4.0 in b6d901fd89ef09545bbfa74e908dc8419c512fd3,
and we no longer install webkit2gtk-4.0 by ourselves unifdef is also no longer necessary